refactor(wifi): Combined AP & STA into single component
