+++ wifi/CMakeLists.txt
@@ -1,4 +1,4 @@
-idf_component_register( SRCS "src/wifi_sta.cpp"
+idf_component_register( SRCS "src/ap.cpp" "src/sta.cpp"
                         INCLUDE_DIRS "include"
                         REQUIRES esp_wifi manager
                         PRIV_REQUIRES esp_event nvs_flash lwip vfs)
