+++ wifi/kconfig
@@ -1,0 +1,22 @@
+menu "Smartknob-HA Components"
+
+    menu "Wifi AP"
+
+        config AP_CHANNEL
+            int "AP Channel to use"
+            default 10
+
+        config AP_MAX_CONNECTIONS
+            int "Max connections to AP"
+            default 2
+
+    endmenu
+
+    menu "Wifi STA"
+
+        config DEFAULT_HOSTNAME
+            string "Default hostname"
+            default "Smartknob"
+    endmenu
+
+endmenu
