--- wifi_ap/kconfig
@@ -1,15 +1,0 @@
-menu "Smartknob-HA Components"
-
-    menu "Wifi AP"
-
-        config AP_CHANNEL
-            int "AP Channel to use"
-            default 10
-
-        config AP_MAX_CONNECTIONS
-            int "Max connections to AP"
-            default 2
-
-    endmenu
-
-endmenu
