--- wifi_sta/component.mk
@@ -1,2 +1,0 @@
-COMPONENT_SRCDIRS:=src
-COMPONENT_ADD_INCLUDEDIRS:=include
