--- wifi_sta/kconfig
@@ -1,11 +1,0 @@
-
-menu "Smartknob-HA Components"
-
-    menu "Wifi STA"
-
-        config DEFAULT_HOSTNAME
-            string "Default hostname"
-            default "Smartknob"
-    endmenu
-
-endmenu
